The logo, searchbox, and buttons are all where they're supposed to be
I'm also taking this opportunity to test, perhaps for the second time, how to input a multi-line commit message for git commit. It seems to be working as I expect it to, honestly.
The changes that I made were achieved through several new CSS rules around the searchbox and the buttons.