Added Footer in the style of Google footer
I added the footer using the same structure as the nav element in the HTML with similar CSS to get the links to justify right.

I also removed the weird center element that I saw in Google's actual homepage and replaced it with CSS on the div element containing the buttons to do the same thing. VS Code didn't even recognize <center> as a tag, that feels very old.